feat(no/sparebankensor): add credit card transaction fetching
https://tinkab.atlassian.net/browse/ITE-3199

+++ src/integration/agents/src/main/java/se/tink/backend/aggregation/agents/nxgen/no/banks/sparebankensor/fetcher/transactionalaccount/entitites/ExecutedEntity.java
@@ -1,13 +1,14 @@
 package se.tink.backend.aggregation.agents.nxgen.no.banks.sparebankensor.fetcher.transactionalaccount.entitites;
 
+import java.math.BigDecimal;
 import se.tink.backend.aggregation.annotations.JsonObject;
 
 @JsonObject
 public class ExecutedEntity {
-    private Double value;
+    private BigDecimal value;
     private String currency;
 
-    public Double getValue() {
+    public BigDecimal getValue() {
         return value;
     }
